Dev commit.

+++ src/com/androidquery/AbstractAQuery.java
@@ -19,10 +19,11 @@ package com.androidquery;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.lang.ref.WeakReference;
 import java.lang.reflect.Constructor;
 import java.nio.channels.FileChannel;
+import java.util.Iterator;
 import java.util.Map;
+import java.util.WeakHashMap;
 
 import android.app.Activity;
 import android.app.Dialog;
@@ -79,8 +80,8 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	private Context context;
 	
 	protected View view;
-	private View progress;
-	private AccountHandle ah;
+	protected View progress;
+	protected AccountHandle ah;
 
 	private T create(View view){
 		
@@ -133,6 +134,8 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 		this.view = root;
 	}
 	
+
+	
 	/**
 	 * Instantiates a new AQuery object.
 	 *
@@ -178,6 +181,25 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 		return create(view);
 	}
 	
+	/**
+	 * Recycle this AQuery object. 
+	 * 
+	 * The method is designed to avoid recreating an AQuery object repeatedly, such as when in list adapter getView method.
+	 *
+	 * @param root The new root of the recycled AQuery.
+	 * @return self
+	 */
+	public T recycle(View root){
+		this.root = root;
+		this.view = null;
+		this.act = null;
+		this.progress = null;
+		this.ah = null;
+		this.context = null;
+		return self();
+	}
+	
+	
 	@SuppressWarnings("unchecked")
 	private T self(){
 		return (T) this;
@@ -201,6 +223,7 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	public T id(int id){
 		view = findView(id);	
 		progress = null;
+		ah = null;
 		return self();
 	}
 	
@@ -213,6 +236,7 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	public T id(View view){
 		this.view = view;	
 		progress = null;
+		ah = null;
 		return self();
 	}
 	
@@ -224,7 +248,8 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	 */
 	public T id(int... path){
 		view = findView(path);	
-		progress = null;		
+		progress = null;
+		ah = null;		
 		return self();
 	}
 	
@@ -262,23 +287,9 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 		return self();
 	}
 	
-	//dequeue the progress, each progress bar only used once per ajax request
-	private View dqProgress(){
-		View result = progress;
-		progress = null;
-		return result;
-	}
-	
 	public T auth(AccountHandle handle){
 		ah = handle;
 		return self();
-	}
-	
-	//dequeue the account handle
-	private AccountHandle dqAccount(){
-		AccountHandle result = ah;
-		ah = null;
-		return result;
 	}
 	
 	
@@ -309,7 +320,28 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 			TextView tv = (TextView) view;
 			tv.setText(text);
 		}
-		
+				
+		return self();
+	}
+	
+	/**
+	 * Set the text of a TextView.
+	 *
+	 * @param text the text
+	 * @return self
+	 */
+	public T text(CharSequence text, boolean gone){
+			
+		if(view instanceof TextView){			
+			TextView tv = (TextView) view;
+			tv.setText(text);
+			if(gone && (text == null || text.length() == 0)){
+				tv.setVisibility(View.GONE);
+			}else{
+				tv.setVisibility(View.VISIBLE);
+			}
+		}
+				
 		return self();
 	}
 	
@@ -509,10 +541,18 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	
 	public T image(String url, boolean memCache, boolean fileCache, int targetWidth, int fallbackId, Bitmap preset, int animId, float ratio){
 		
+		/*
 		BitmapAjaxCallback cb = new BitmapAjaxCallback();		
-		cb.url(url).memCache(memCache).fileCache(fileCache).targetWidth(targetWidth).fallback(fallbackId).preset(preset).animation(animId).ratio(ratio);	
-		
+		cb.url(url).memCache(memCache).fileCache(fileCache).targetWidth(targetWidth).fallback(fallbackId).preset(preset).animation(animId).ratio(ratio);			
 		return image(cb);
+		*/
+		
+		if(view instanceof ImageView){		
+			BitmapAjaxCallback.async(getContext(), (ImageView) view, url, memCache, fileCache, targetWidth, fallbackId, preset, animId, ratio, progress);
+			progress = null;
+		}
+		
+		return self();
 	}
 	
 	
@@ -527,8 +567,13 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	
 	public T image(BitmapAjaxCallback callback){
 		
-		if(view instanceof ImageView || view instanceof TextView){			
-			callback.view(view).progress(dqProgress()).async(getContext());			
+		if(view instanceof ImageView){			
+			callback.imageView((ImageView) view);
+			if(progress != null){
+				callback.progress(progress);
+				progress = null;
+			}
+			callback.async(getContext());						
 		} 
 		
 		return self();
@@ -611,6 +656,36 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 		return image(cb);
 	}
 	
+	/**
+	 * Set tag object of a view.
+	 *
+	 * @param tag 
+	 * @return self
+	 */
+	public T tag(Object tag){
+		
+		if(view != null){
+			view.setTag(tag);
+		}
+		
+		return self();
+	}
+	
+	/**
+	 * Set tag object of a view.
+	 *
+	 * @param key
+	 * @param tag 
+	 * @return self
+	 */
+	public T tag(int key, Object tag){
+		
+		if(view != null){
+			view.setTag(key, tag);
+		}
+		
+		return self();
+	}
 	
 	/**
 	 * Set a view to be transparent.
@@ -750,6 +825,21 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 				view.setBackgroundDrawable(null);
 			}
 		
+		}
+		
+		return self();
+	}
+	
+	/**
+	 * Set view background color.
+	 *
+	 * @param color
+	 * @return self
+	 */
+	public T backgroundColor(int color){
+		
+		if(view != null){		
+			view.setBackgroundColor(color);		
 		}
 		
 		return self();
@@ -1286,33 +1376,60 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	}
 	
 	/**
-	 * Set the width of a view. Notes all parameters are in DIP, not in pixel. 
-	 * Input can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
+	 * Set the width of a view in dip. 
+	 * Can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
 	 *
 	 * @param dip width in dip
 	 * @return self
 	 */
 	
 	public T width(int dip){		
-		size(true, dip);		
+		size(true, dip, true);		
 		return self();
 	}
 	
 	/**
-	 * Set the height of a view. Notes all parameters are in DIP, not in pixel. 
-	 * Input can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
+	 * Set the height of a view in dip. 
+	 * Can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
 	 *
 	 * @param dip height in dip
 	 * @return self
 	 */
 	
 	public T height(int dip){		
-		size(false, dip);
+		size(false, dip, true);
 		return self();
 	}
 	
+	/**
+	 * Set the width of a view in dip or pixel. 
+	 * Can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
+	 *
+	 * @param width width
+	 * @param dip dip or pixel
+	 * @return self
+	 */
 	
-	private void size(boolean width, int dip){
+	public T width(int width, boolean dip){		
+		size(true, width, dip);		
+		return self();
+	}
+	
+	/**
+	 * Set the height of a view in dip or pixel. 
+	 * Can also be ViewGroup.LayoutParams.FILL_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, or ViewGroup.LayoutParams.MATCH_PARENT.
+	 *
+	 * @param height height
+	 * @param dip dip or pixel
+	 * @return self
+	 */
+	
+	public T height(int height, boolean dip){		
+		size(false, height, dip);
+		return self();
+	}
+	
+	private void size(boolean width, int n, boolean dip){
 		
 		if(view != null){
 		
@@ -1320,14 +1437,14 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 			
 			Context context = getContext();
 				
-			if(dip > 0){
-				dip = AQUtility.dip2pixel(context, dip);
+			if(n > 0 && dip){
+				n = AQUtility.dip2pixel(context, n);
 			}
 			
 			if(width){
-				lp.width = dip;
+				lp.width = n;
 			}else{
-				lp.height = dip;
+				lp.height = n;
 			}
 			
 			view.setLayoutParams(lp);
@@ -1366,7 +1483,18 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	
 	public <K> T ajax(AjaxCallback<K> callback){
 				
-		callback.progress(dqProgress()).auth(dqAccount()).async(getContext());
+		
+		if(ah != null){
+			callback.auth(ah);
+			ah = null;
+		}
+		
+		if(progress != null){
+			callback.progress(progress);
+			progress = null;
+		}
+		
+		callback.async(getContext());
 		
 		return self();
 	}	
@@ -1561,8 +1689,7 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	public T invalidate(String url){
 		
 		File file = getCachedFile(url);
-		file.delete();
-		
+		if(file != null) file.delete();		
 		return self();
 	}
 	
@@ -1812,17 +1939,29 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	}
 	
 	
-	private static WeakReference<Dialog> diaRef;
+	//weak hash map that holds the dialogs so they will never memory leaked
+	private static WeakHashMap<Dialog, Void> dialogs = new WeakHashMap<Dialog, Void>();
 	
 	public T show(Dialog dialog){
 		
-		if(diaRef != null && diaRef.get() != dialog){
-			dismiss();
+		try{
+			if(dialog != null){
+				dialog.show();
+				dialogs.put(dialog, null);
+			}
+		}catch(Exception e){			
 		}
 		
+		return self();
+	}
+	
+	public T dismiss(Dialog dialog){
+		
 		try{
-			dialog.show();
-			diaRef = new WeakReference<Dialog>(dialog);
+			if(dialog != null){			
+				dialogs.remove(dialog);
+				dialog.dismiss();
+			}
 		}catch(Exception e){			
 		}
 		
@@ -1831,17 +1970,18 @@ public abstract class AbstractAQuery<T extends AbstractAQuery<T>> implements Con
 	
 	public T dismiss(){
 		
-		if(diaRef != null){
-			Dialog d = diaRef.get();
-			if(d != null && d.isShowing()){
-				try{
-					d.dismiss();
-				}catch(Exception e){					
-				}
-			}
-			diaRef = null;
-		}
+		Iterator<Dialog> keys = dialogs.keySet().iterator();
 		
+		while(keys.hasNext()){
+			
+			Dialog d = keys.next();
+			try{
+				d.dismiss();
+			}catch(Exception e){				
+			}
+			keys.remove();
+			
+		}
 		return self();
 		
 	}
